+++ src/Console/Application.php
@@ -7,8 +7,11 @@ use Castor\Console\Command\TaskCommand;
 use Castor\Context;
 use Castor\ContextDescriptor;
 use Castor\ContextRegistry;
+use Castor\Event\AfterApplicationInitializationEvent;
+use Castor\EventDispatcher;
 use Castor\FunctionFinder;
 use Castor\GlobalHelper;
+use Castor\ListenerDescriptor;
 use Castor\Monolog\Processor\ProcessProcessor;
 use Castor\PlatformUtil;
 use Castor\SectionOutput;
@@ -25,6 +28,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\VarDumper\Cloner\AbstractCloner;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\HttpClient\Exception\ExceptionInterface as HttpExceptionInterface;
 
@@ -43,12 +47,16 @@ class Application extends SymfonyApplication
         private readonly ContextRegistry $contextRegistry = new ContextRegistry(),
         private readonly StubsGenerator $stubsGenerator = new StubsGenerator(),
         private readonly FunctionFinder $functionFinder = new FunctionFinder(),
+        private readonly EventDispatcher $eventDispatcher = new EventDispatcher(),
     ) {
         if (!class_exists(\RepackedApplication::class)) {
             $this->add(new RepackCommand());
         }
 
         $this->setCatchErrors(true);
+
+        AbstractCloner::$defaultCasters[self::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
+        AbstractCloner::$defaultCasters[AfterApplicationInitializationEvent::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
 
         parent::__construct(static::NAME, static::VERSION);
     }
@@ -60,6 +68,7 @@ class Application extends SymfonyApplication
         $sectionOutput = new SectionOutput($output);
 
         GlobalHelper::setApplication($this);
+        GlobalHelper::setEventDispatcher($this->eventDispatcher);
         GlobalHelper::setInput($input);
         GlobalHelper::setSectionOutput($sectionOutput);
         GlobalHelper::setLogger(new Logger(
@@ -114,6 +123,12 @@ class Application extends SymfonyApplication
                 $tasks[] = $function;
             } elseif ($function instanceof ContextDescriptor) {
                 $this->contextRegistry->add($function);
+            } elseif ($function instanceof ListenerDescriptor && null !== $function->reflectionFunction->getClosure()) {
+                $this->eventDispatcher->addListener(
+                    $function->asListener->event,
+                    $function->reflectionFunction->getClosure(),
+                    $function->asListener->priority
+                );
             }
         }
 
@@ -133,6 +148,8 @@ class Application extends SymfonyApplication
                 $contextNames,
             ));
         }
+
+        $this->eventDispatcher->dispatch(new AfterApplicationInitializationEvent($this, $tasks));
 
         return $tasks;
     }