+++ src/Event/AfterExecuteTaskEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Castor\Event;
+
+use Castor\Console\Command\TaskCommand;
+
+class AfterExecuteTaskEvent
+{
+    public function __construct(
+        public readonly TaskCommand $task,
+        public readonly mixed $result,
+    ) {
+    }
+}