+++ src/Event/BeforeExecuteTaskEvent.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Castor\Event;
+
+use Castor\Console\Command\TaskCommand;
+
+class BeforeExecuteTaskEvent
+{
+    public function __construct(
+        public readonly TaskCommand $task,
+    ) {
+    }
+}