+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released yet
 
+* Add `AsListener` attribute to register a task as event listener
 * Add `force` argument to `fingerprint()` method to force run the callable, even if fingerprint is same
 * Allow to override `AsTask` and `AsContext` attributes
 * Add `wait_for()`, `wait_for_port()`, `wait_for_url()`, `wait_for_http_status()` functions